# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iApp API Documentation",
        "description": "API Documentation - iApp",
        "contact": {
            "email": "akash.corp@akij.net"
        },
        "license": {
            "name": "@iapp.akij.net",
            "url": "https://iapp.akij.net"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logged in to the system by user",
                "description": "Logged in to the system by user",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "akash@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login User Data, with Token"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            },
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}